# Adds a parallell gzip library