--- a/artifact/compressor_gzip_test.go
+++ b/artifact/compressor_gzip_test.go
@@ -40,7 +40,14 @@ func TestCompressorGzip(t *testing.T) {
 
 	rbuf := make([]byte, len(testData))
 	i, err = r.Read(rbuf)
-	assert.Equal(t, err, io.EOF)
 	assert.Equal(t, i, len(testData))
+	// It is perfectly valid for a reader to return nil on an exhaustive
+	// write when n-bytes read > 0
+	assert.True(t, err == io.EOF  || err == nil)
 	assert.Equal(t, []byte(testData), rbuf)
+
+	// Second read must return zero bytes, and EOF though
+	i, err = r.Read(rbuf)
+	assert.Equal(t, err, io.EOF)
+	assert.Equal(t, i, 0)
 }